+++ FSharp/MirrorSharpFSharpOptions.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using FSharp.Compiler.CodeAnalysis;
 using Microsoft.FSharp.Core;