+++ IL/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+- Target setting on IILSession
+
 ## [0.1] - 2021-08-31
 
 ### Added